Compare-methods {mondate}	R Documentation

Comparison Methods

Description

Methods for the Compare group of functions.

Methods

signature(e1 = "mondate", e2 = "mondate")

compares two mondates. The usual recycling rules apply to the shorter of the two mondates. The result will be logical. The usual rules apply as to the shape of the result.

See Also

Compare

Examples

A<-mondate.ymd(2001:2003,12,31) # three year ends
B<-mondate.ymd(2001:2003, 6,30) # three mid-years
B<A                             # c(TRUE, TRUE, TRUE)

[Package mondate version 0.10.02 Index]