Summary-methods {mondate}	R Documentation

Summary Methods

Description

Methods for the Summary group of functions.

Methods

signature(x = "mondate")

summarizes a mondate.

The result will be a mondate with the same displayFormat and timeunits properties.

The usual rules apply as to the shape of the result.

Examples

A<-mondate.ymd(2001:2010,12,31) # ten yearends
min(A)                  # December 31, 2001
max(A)                  # December 31, 2010
range(A)

[Package mondate version 0.10.02 Index]