YearQuartersFormat {mondate}	R Documentation

Formatting Functions for mondate Objects

Description

Functions to format a mondate into its character representation according to the displayFormat property.

Usage

YearQuartersFormat(x)

Arguments

x
a mondate or a Date or a POSIXt.

Details

YearQuartersFormat is an example of a special formatting function that can be provided to a mondate object when created. It will represent the date as YYYYQ* where * is 1-4. See Examples

Examples

b <- mondate(1:12, formatFUN = YearQuartersFormat)   # end of first 12 months of 2000
b         

[Package mondate version 0.10.02 Index]