as.mondate {mondate}	R Documentation

As.Mondate Method

Description

Coerce an object to class mondate.

Usage

as.mondate(x, ...)

Arguments

x
an R object.

...
optional arguments passed to other methods.

Details

This is a convenience function that simply calls the appropriate mondate conversion method depending on the class of x.

Value

A mondate if coersion is successful.

Author(s)

Dan Murphy

See Also

mondate-methods

Examples

y <- as.Date("2012-12-31")
as.mondate(y)

[Package mondate version 0.10.02 Index]