diff.mondate {mondate}	R Documentation

'diff' for package mondate

Description

Returns suitably lagged and iterated differences of an object of class mondate.

Usage

## S3 method for class 'mondate'
diff(x, lag = 1L, differences = 1L, ...)

Arguments

x
a mondate vector or matrix containing the values to be differenced.

lag
an integer indicating which lag to use.

differences
an integer indicating the order of the difference.

...
further arguments to be passed to or from methods.

Details

See the diff function in base.

Value

If x is a vector of length n and differences=1, then the computed result is equal to the successive differences x[(1+lag):n] - x[1:(n-lag)].

If difference is larger than one this algorithm is applied recursively to x. Note that the returned value is a vector which is shorter than x.

If x is a matrix then the difference operations are carried out on each column separately.

Author(s)

Dan Murphy

See Also

diff

Examples

evalDate<-mondate(c(12+12*1:10))  # year ends 2001, ..., 2010
diff(evalDate)                    # vector of length 9 holding the number 12, 
                                  # with "timeunits" attribute = "months"

[Package mondate version 0.10.02 Index]