print-methods {mondate}	R Documentation

Methods to Display a Mondate

Description

Methods to display a mondate in an R session.

Usage

    ## S3 method for class 'mondate'
print(x, ...)
    ## S4 method for signature 'mondate'
show(object)

Arguments

x
a mondate

object
a mondate

...
arguments passed to and from other methods

Methods

print(x = "mondate", ...)

Print the date in its character representation using the
displayFormat property. Uses the noquote function. Arguments in ... are passed to base::print.

show(object = "mondate")

Same as print above, but with a "header" showing the timeunits property.

[Package mondate version 0.10.02 Index]