timeunits-methods {mondate}	R Documentation

Methods to Access 'timeunits' Property

Description

Methods to get and set the timeunits value of a mondate.

Usage

## S4 method for signature 'mondate'
timeunits(x)
## S4 method for signature 'ANY'
timeunits(x)
## S4 replacement method for signature 'mondate'
timeunits(x)<-value

Arguments

x
a mondate.

value
For the "get" method, a character string indicating the units with which to measure time as a result of operations on a mondate. Choices are currently

"months"

"years"

"days"

If x is not a mondate, the "get" value returned is NULL.

Note

The mondateTimeunits versions have been deprecated.

Examples

x <- mondate("2010-6-30") # The middle of 2010
timeunits(x)              # "months"
y <- x + 12               # One year (12 months) later.
timeunits(y)              # "months"
y - x                     # Result is 12 months
timeunits(y) <- "years"
y - x                     # Result is 1 year, with a warning because 
                          # x@timeunits != y@timeunits. Units of result
                          # comes from the first argument.
timeunits(y) <- "days"
suppressWarnings(y - x)   # Result is 365 days -- warning suppressed

[Package mondate version 0.10.02 Index]