[-methods {mondate}	R Documentation

Extraction Method for Mondates

Description

Methods to extract portions of a mondate

Usage

## S3 method for class 'mondate'
head(x, ...)
## S3 method for class 'mondate'
tail(x, ...)

Arguments

x
a mondate.

...
See the base functions for details.

Details

Whether the mondate x is shaped as a vector or a matrix, the head and rbind methods will behave just as they would if x were numeric.

Methods

[(x = "mondate")

See [ for more details.

Examples

(m<-structure(mondate.ymd(2001:2010,12,31),names=LETTERS[1:10]))
m[1]
m[2:5]
head(m)
tail(m,2)

(M<-cbind(m-12,m,m+12, deparse.level=2))  # a matrix
M[1:5,1:2]    # '[' works with matrix mondates
head(M,2)  # as does 'head'

[Package mondate version 0.10.02 Index]